% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{EA_set_env}
\alias{EA_set_env}
\title{Single exon transcripts contained in this package's transcript set}
\description{
Single exon transcripts contained in this package's transcript set
}
\details{
Envirnonment, with transcript names *of single exon transcripts only* as keys, and \code{TRUE} as values.
}
\author{
Dennis Kostka
}
\keyword{data}
